[Translator] merged xlif classes
