--- src/Symfony/Component/Translation/Tests/Loader/XliffFileLoaderWithResnameTest.php
@@ -1,30 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\Translation\Tests\Loader;
-
-use Symfony\Component\Translation\Loader\XliffFileWithResnameLoader;
-
-class XliffFileWithResnameLoaderTest extends XliffFileLoaderTest
-{
-    public function testLoadWithResname()
-    {
-        $loader = $this->createLoader();
-        $catalogue = $loader->load(__DIR__.'/../fixtures/resname.xlf', 'en', 'domain1');
-
-        $this->assertEquals(array('foo' => 'bar', 'bar' => 'baz', 'baz' => 'foo'), $catalogue->all('domain1'));
-    }
-
-    protected function createLoader()
-    {
-        return new XliffFileWithResnameLoader();
-    }
-}
