--- src/Symfony/Component/Translation/Tests/fixtures/resname-clean.xlf
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<xliff xmlns="urn:oasis:names:tc:xliff:document:1.2" version="1.2">
-  <file source-language="en" datatype="plaintext" original="file.ext">
-    <body>
-      <trans-unit id="1" resname="foo">
-        <source>foo</source>
-        <target>bar</target>
-      </trans-unit>
-      <trans-unit id="2" resname="key">
-        <source>key</source>
-        <target></target>
-      </trans-unit>
-    </body>
-  </file>
-</xliff>
